Renamed icon to CopyLinkIcon
