+++ src/lib/components/icons/copy-link.svelte
